requirements.txt
asgiref                       
certifi                       
cffi                          
chardet             
coreapi                    
coreschema                   
cryptography                
defusedxml                  
Django                     
django-templated-mail       
djangorestframework         
djangorestframework-simplejwt 
djoser                      
idna                        
itypes                     
Jinja2                      
MarkupSafe                
oauthlib                   
pip                        
pkg-resources                
psycopg2                      
pycparser                     
PyJWT                           
python3-openid                 
pytz                        
requests                      
requests-oauthlib           
setuptools                   
six                         
social-auth-app-django       
social-auth-core              
sqlparse                  
uritemplate                  
urllib3                        
wheel                         